Safari-609.1.11.1/LayoutTests/inspector/canvas/recording-expected.txt

Test general cases of CanvasAgent recording calls.


== Running test suite: Canvas.recording
-- Running test case: Canvas.ActionParameterNaN
PASS: The recording should have 1 frame.
PASS: The first frame should have 1 action.
PASS: The action should have 1 parameter.
PASS: The parameter should be null.

-- Running test case: Canvas.MultipleRecording
Starting a recording of canvas 1...
PASS: Recording started of canvas 1
Starting a recording of canvas 2...
PASS: Recording started of canvas 2
Performing actions...
PASS: Actions performed.
Stopping the recording of canvas 1...
PASS: There should be a recording for canvas 1.
Stopping the recording of canvas 2...
PASS: There should be a recording for canvas 2.

-- Running test case: Canvas.NoActions
PASS: A recording should have been started and stopped once.
PASS: A recording should have been started and stopped twice.

-- Running test case: Canvas.startRecording.InvalidCanvasId
PASS: Should produce an error.
Error: Missing canvas for given canvasId

-- Running test case: Canvas.stopRecording.InvalidCanvasId
PASS: Should produce an error.
Error: Missing canvas for given canvasId
